pyproject.toml
[project]
name = "easwee-net"
version = "4.0.0"
description = "Easwee's digital playground"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "fastapi[standard]>=0.115.6",
    "jinja-markdown>=1.210911",
    "llama-index-embeddings-openai>=0.3.1",
    "llama-index-graph-stores-neo4j>=0.4.4",
    "llama-index-llms-anthropic>=0.6.3",
    "llama-index-llms-gemini>=0.4.2",
    "llama-index-llms-mistralai>=0.3.1",
    "llama-index-llms-openai-like>=0.3.3",
    "llama-index-llms-openai>=0.3.12",
    "markdown2>=2.5.2",
    "openai>=1.58.1",
    "python-dotenv>=1.0.1",
    "pandas>=2.2.3",
    "pyarrow>=19.0.0",
]
